fix: fixed by renaming to .jsx
